Move tests to match core code.
